END TO END YOUTUBE TRANSCRIBE SUMMARIZE LLM APP/requirements.txt
google-generativeai
python-dotenv
youtube_transcript_api
pathlib
streamlit 